+++ Makefile
@@ -60,7 +60,6 @@ mypy:
 .PHONY: test
 test:
 	pytest --cov=pydantic
-	@python tests/try_assert.py
 
 .PHONY: testcov
 testcov: test
